// src/enums/IPCEnum.ts
export enum IPCEnum {
    PLAY = "Play",
    LAUNCHER_CLOSED = "LauncherClosed",
    INSTALL = "Install",
    NEW_STEP = "NewStep",
    DOWNLOAD_STARTED = "DownloadStarted",
    DOWNLOAD_PROGRESSED = "DownloadProgressed",
    DOWNLOAD_FINISHED = "DownloadFinished",
    INSTALLATION_FINISHED = "InstallationFinished"
}